requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-dotenv>=1.0.0
cohere>=5.20.0
qdrant-client>=1.16.1
openai-agents>=0.6.2
requests>=2.32.5
trafilatura>=2.0.0
pydantic>=2.9.0